requirement.txt
bs4
kafka-python
flask
google-cloud-storage
google-cloud-bigquery-storage
pyarrow
pandas
kubernetes